# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/pedidos": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [
          {
            "name": "x-correlation-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/OrderResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Pedidos"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "pagina",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limite",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "busca",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "type": "object",
                      "required": [
                        "itens",
                        "total",
                        "pagina",
                        "limite"
                      ],
                      "properties": {
                        "itens": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pagina": {
                          "type": "integer"
                        },
                        "limite": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Pedidos"
        ]
      }
    },
    "/api/v1/pedidos/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/OrderResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Pedidos"
        ]
      }
    },
    "/api/v1/pedidos/{id}/eventos": {
      "get": {
        "operationId": "OrdersController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pagina",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limite",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "type": "object",
                      "required": [
                        "itens",
                        "total",
                        "pagina",
                        "limite"
                      ],
                      "properties": {
                        "itens": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/HistoryResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pagina": {
                          "type": "integer"
                        },
                        "limite": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Pedidos"
        ]
      }
    },
    "/api/v1/etiquetas": {
      "post": {
        "operationId": "TagsController_register",
        "parameters": [
          {
            "name": "x-correlation-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/ProvisioningResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Etiquetas"
        ]
      }
    },
    "/api/v1/etiquetas/{uid}": {
      "get": {
        "operationId": "TagsController_get",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/ProvisioningResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Etiquetas"
        ]
      }
    },
    "/api/v1/provisionamentos/{id}": {
      "get": {
        "operationId": "ProvisioningsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/ProvisioningResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Provisionamentos"
        ]
      }
    },
    "/api/v1/provisionamentos/{id}/ativacao": {
      "post": {
        "operationId": "ProvisioningsController_activateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/ProvisioningResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Provisionamentos"
        ]
      }
    },
    "/api/v1/provisionamentos/{id}/encerramento": {
      "post": {
        "operationId": "ProvisioningsController_closeTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sucesso",
                    "mensagem",
                    "dados"
                  ],
                  "properties": {
                    "sucesso": {
                      "type": "boolean",
                      "example": true
                    },
                    "mensagem": {
                      "type": "string"
                    },
                    "dados": {
                      "$ref": "#/components/schemas/ProvisioningResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Provisionamentos"
        ]
      }
    },
...
  },
  "info": {
    "title": "NFC Trace API",
    "description": "API experimental sem login. Operador e dispositivo são declarados. HTTP 200 em eventos confirma armazenamento; a decisão informa se a movimentação foi autorizada. UID/NDEF não são autenticação criptográfica.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "OrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "string"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "enum": [
              "CADASTRADO",
              "COLETADO",
              "RECEBIDO",
              "ENTREGUE"
            ]
          },
          "expedido": {
            "type": "boolean"
          },
          "versao": {
            "type": "number"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "codigo",
          "descricao",
          "estado",
          "expedido",
          "versao",
          "criadoEm"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "sucesso": {
            "type": "boolean",
            "example": false
          },
          "mensagem": {
            "type": "string"
          },
          "codigo": {
            "type": "string",
            "example": "IDEMPOTENCIA_CONFLITO"
          },
          "correlacaoId": {
            "type": "string"
          }
        },
        "required": [
          "sucesso",
          "mensagem",
          "codigo",
          "correlacaoId"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string",
            "example": "TCC-001",
            "maxLength": 64
          },
          "descricao": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "codigo"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}